feat: 포커싱 관련 로직 useFocus로 분리
